[kernel] B: Fix x86_64 Integration Tests
Four issues prevented Rust integration tests from passing on x86_64:

1. TLS inline assembly: test-kernel used %gs for TDA access (x86); x86_64
   uses %fs via FS_BASE MSR. Added #[cfg(target_arch)] guards in
   read_gs_offset_0() and read_gs_at() for architecture-specific assembly.

2. MMIO RAMFS page fault: kctrl() only updated software bookkeeping
   tables, not the x86_64 hardware page tables (4-level PML4 hierarchy).
   Added hwpt::map() call in kctrl() for x86_64, and updated
   ensure_table()/map() in hwpt.rs to propagate PTE_USER through all
   page table levels when user access is required.

3. Worker thread RIP=0 crash: _do_start_thread (x86_64) read func/arg
   from RDX/RCX, but __leave_kernel_to_user_mode passes them in RDI/RSI.
   Fixed the assembly stub to use the correct registers.

4. Mutex/condvar timeout rejection: Kernel checked timeout sentinels
   against usize::MAX, but user-space sends u32::MAX. On x86_64,
   u32::MAX != usize::MAX, causing infinite-timeout requests to fail with
   "invalid timeout". Changed sentinel checks to use u32::MAX as usize.

All 7 Rust integration tests now pass on both x86 and x86_64:
  test-kernel, arch-rust, thread-rust, stress-rust, file-rust,
  linux-app, vfs-test.

+++ src/kernel/src/pm/kcall/lock_mutex.rs
@@ -73,7 +73,9 @@ pub unsafe fn lock_mutex(
     );
     // Unpack kernel call arguments.
     let mutex_addr: MutexAddress = MutexAddress::from(mutex_addr);
-    let timeout: Option<SystemTime> = if timeout_s == usize::MAX && timeout_ns == usize::MAX {
+    let timeout: Option<SystemTime> = if timeout_s == u32::MAX as usize
+        && timeout_ns == u32::MAX as usize
+    {
         None
     } else {
         match SystemTime::new(timeout_s as u64, timeout_ns as u32) {
